// web/src/api/models/ResponsesGithubWidgetCommitsDayData.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ResponsesGithubWidgetCommitsDayData
 */
export interface ResponsesGithubWidgetCommitsDayData {
    /**
     * 
     * @type {string}
     * @memberof ResponsesGithubWidgetCommitsDayData
     */
    color?: string;
    /**
     * 
     * @type {number}
     * @memberof ResponsesGithubWidgetCommitsDayData
     */
    count?: number;
    /**
     * 
     * @type {string}
     * @memberof ResponsesGithubWidgetCommitsDayData
     */
    date?: string;
    /**
     * 
     * @type {number}
     * @memberof ResponsesGithubWidgetCommitsDayData
     */
    percent?: number;
}

/**
 * Check if a given object implements the ResponsesGithubWidgetCommitsDayData interface.
 */
export function instanceOfResponsesGithubWidgetCommitsDayData(value: object): value is ResponsesGithubWidgetCommitsDayData {
    return true;
}

export function ResponsesGithubWidgetCommitsDayDataFromJSON(json: any): ResponsesGithubWidgetCommitsDayData {
    return ResponsesGithubWidgetCommitsDayDataFromJSONTyped(json, false);
}

export function ResponsesGithubWidgetCommitsDayDataFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResponsesGithubWidgetCommitsDayData {
    if (json == null) {
        return json;
    }
    return {
        
        'color': json['color'] == null ? undefined : json['color'],
        'count': json['count'] == null ? undefined : json['count'],
        'date': json['date'] == null ? undefined : json['date'],
        'percent': json['percent'] == null ? undefined : json['percent'],
    };
}

export function ResponsesGithubWidgetCommitsDayDataToJSON(json: any): ResponsesGithubWidgetCommitsDayData {
    return ResponsesGithubWidgetCommitsDayDataToJSONTyped(json, false);
}

export function ResponsesGithubWidgetCommitsDayDataToJSONTyped(value?: ResponsesGithubWidgetCommitsDayData | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'color': value['color'],
        'count': value['count'],
        'date': value['date'],
        'percent': value['percent'],
    };
}