// web/src/api/models/ResponsesGithubWidgetCommitsWeekData.ts
export interface ResponsesGithubWidgetCommitsWeekData {
	/**
	 *
	 * @type {ResponsesGithubWidgetCommitsDayData}
	 * @memberof ResponsesGithubWidgetCommitsWeekData
	 */
	friday?: ResponsesGithubWidgetCommitsDayData;
	/**
	 *
	 * @type {ResponsesGithubWidgetCommitsDayData}
	 * @memberof ResponsesGithubWidgetCommitsWeekData
	 */
	monday?: ResponsesGithubWidgetCommitsDayData;
	/**
	 *
	 * @type {ResponsesGithubWidgetCommitsDayData}
	 * @memberof ResponsesGithubWidgetCommitsWeekData
	 */
	saturday?: ResponsesGithubWidgetCommitsDayData;
	/**
	 *
	 * @type {ResponsesGithubWidgetCommitsDayData}
	 * @memberof ResponsesGithubWidgetCommitsWeekData
	 */
	sunday?: ResponsesGithubWidgetCommitsDayData;
	/**
	 *
	 * @type {ResponsesGithubWidgetCommitsDayData}
	 * @memberof ResponsesGithubWidgetCommitsWeekData
	 */
	thursday?: ResponsesGithubWidgetCommitsDayData;
	/**
	 *
	 * @type {ResponsesGithubWidgetCommitsDayData}
	 * @memberof ResponsesGithubWidgetCommitsWeekData
	 */
	tuesday?: ResponsesGithubWidgetCommitsDayData;
	/**
	 *
	 * @type {ResponsesGithubWidgetCommitsDayData}
	 * @memberof ResponsesGithubWidgetCommitsWeekData
	 */
	wednesday?: ResponsesGithubWidgetCommitsDayData;
	/**
	 *
	 * @type {string}
	 * @memberof ResponsesGithubWidgetCommitsWeekData
	 */
	weekStart?: string;
}