// web/src/api/models/ResponsesGithubWidgetCommitsData.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ResponsesGithubWidgetCommitsWeekData } from './ResponsesGithubWidgetCommitsWeekData';
import {
	ResponsesGithubWidgetCommitsWeekDataFromJSON,
	ResponsesGithubWidgetCommitsWeekDataFromJSONTyped,
	ResponsesGithubWidgetCommitsWeekDataToJSON,
	ResponsesGithubWidgetCommitsWeekDataToJSONTyped,
} from './ResponsesGithubWidgetCommitsWeekData';

/**
 *
 * @export
 * @interface ResponsesGithubWidgetCommitsData
 */
export interface ResponsesGithubWidgetCommitsData {
	/**
	 *
	 * @type {number}
	 * @memberof ResponsesGithubWidgetCommitsData
	 */
	total?: number;
	/**
	 *
	 * @type {Array<ResponsesGithubWidgetCommitsWeekData>}
	 * @memberof ResponsesGithubWidgetCommitsData
	 */
	weeks?: Array<ResponsesGithubWidgetCommitsWeekData>;
}

/**
 * Check if a given object implements the ResponsesGithubWidgetCommitsData interface.
 */
export function instanceOfResponsesGithubWidgetCommitsData(
	value: object,
): value is ResponsesGithubWidgetCommitsData {
	return true;
}

export function ResponsesGithubWidgetCommitsDataFromJSON(
	json: any,
): ResponsesGithubWidgetCommitsData {
	return ResponsesGithubWidgetCommitsDataFromJSONTyped(json, false);
}

export function ResponsesGithubWidgetCommitsDataFromJSONTyped(
	json: any,
	ignoreDiscriminator: boolean,
): ResponsesGithubWidgetCommitsData {
	if (json == null) {
		return json;
	}
	return {
		total: json['total'] == null ? undefined : json['total'],
		weeks:
			json['weeks'] == null
				? undefined
				: (json['weeks'] as Array<any>).map(
						ResponsesGithubWidgetCommitsWeekDataFromJSON,
					),
	};
}

export function ResponsesGithubWidgetCommitsDataToJSON(
	json: any,
): ResponsesGithubWidgetCommitsData {
	return ResponsesGithubWidgetCommitsDataToJSONTyped(json, false);
}

export function ResponsesGithubWidgetCommitsDataToJSONTyped(
	value?: ResponsesGithubWidgetCommitsData | null,
	ignoreDiscriminator: boolean = false,
): any {
	if (value == null) {
		return value;
	}

	return {
		total: value['total'],
		weeks:
			value['weeks'] == null
				? undefined
				: (value['weeks'] as Array<any>).map(
						ResponsesGithubWidgetCommitsWeekDataToJSON,
					),
	};
}