// web/src/api/models/ResponsesGithubWidgetProfileData.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ResponsesGithubWidgetProfileData
 */
export interface ResponsesGithubWidgetProfileData {
    /**
     * 
     * @type {string}
     * @memberof ResponsesGithubWidgetProfileData
     */
    avatarUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof ResponsesGithubWidgetProfileData
     */
    bio?: string;
    /**
     * 
     * @type {string}
     * @memberof ResponsesGithubWidgetProfileData
     */
    company?: string;
    /**
     * 
     * @type {number}
     * @memberof ResponsesGithubWidgetProfileData
     */
    followers?: number;
    /**
     * 
     * @type {number}
     * @memberof ResponsesGithubWidgetProfileData
     */
    following?: number;
    /**
     * 
     * @type {string}
     * @memberof ResponsesGithubWidgetProfileData
     */
    htmlUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof ResponsesGithubWidgetProfileData
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof ResponsesGithubWidgetProfileData
     */
    publicGists?: number;
    /**
     * 
     * @type {number}
     * @memberof ResponsesGithubWidgetProfileData
     */
    publicRepos?: number;
}

/**
 * Check if a given object implements the ResponsesGithubWidgetProfileData interface.
 */
export function instanceOfResponsesGithubWidgetProfileData(value: object): value is ResponsesGithubWidgetProfileData {
    return true;
}

export function ResponsesGithubWidgetProfileDataFromJSON(json: any): ResponsesGithubWidgetProfileData {
    return ResponsesGithubWidgetProfileDataFromJSONTyped(json, false);
}

export function ResponsesGithubWidgetProfileDataFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResponsesGithubWidgetProfileData {
    if (json == null) {
        return json;
    }
    return {
        
        'avatarUrl': json['avatar_url'] == null ? undefined : json['avatar_url'],
        'bio': json['bio'] == null ? undefined : json['bio'],
        'company': json['company'] == null ? undefined : json['company'],
        'followers': json['followers'] == null ? undefined : json['followers'],
        'following': json['following'] == null ? undefined : json['following'],
        'htmlUrl': json['html_url'] == null ? undefined : json['html_url'],
        'name': json['name'] == null ? undefined : json['name'],
        'publicGists': json['public_gists'] == null ? undefined : json['public_gists'],
        'publicRepos': json['public_repos'] == null ? undefined : json['public_repos'],
    };
}

export function ResponsesGithubWidgetProfileDataToJSON(json: any): ResponsesGithubWidgetProfileData {
    return ResponsesGithubWidgetProfileDataToJSONTyped(json, false);
}

export function ResponsesGithubWidgetProfileDataToJSONTyped(value?: ResponsesGithubWidgetProfileData | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'avatar_url': value['avatarUrl'],
        'bio': value['bio'],
        'company': value['company'],
        'followers': value['followers'],
        'following': value['following'],
        'html_url': value['htmlUrl'],
        'name': value['name'],
        'public_gists': value['publicGists'],
        'public_repos': value['publicRepos'],
    };
}